Debugging Q2

+++ src/main/scala/ch/ethz/queries/Q4.scala
@@ -51,7 +51,7 @@ class Q4 extends ChQuery{
         r.getField("OL_SUPPLY_W_ID").asInstanceOf[Int],
         r.getField("OL_DELIVERY_D").asInstanceOf[Long],
         r.getField("OL_QUANTITY").asInstanceOf[Short],
-        r.getField("OL_AMOUNT").asInstanceOf[Double],
+        r.getField("OL_AMOUNT").asInstanceOf[Long],
         r.getField("OL_DIST_INFO").asInstanceOf[String]
       )
     }).toDF()
